fix(client/registerForm): move the validation file to utils and update de imports
