+++ client/src/utils/registerValidation.ts
